% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\docType{data}
\name{.central}
\alias{.central}
\title{Central Zonal Council}
\format{
An object of class \code{character} of length 4.
}
\usage{
.central
}
\description{
Central Zonal Council comprises of Chhattisgarh, Madhya Pradesh,
Uttarakhand and Uttar Pradesh.
}
\details{
See
\url{https://en.wikipedia.org/wiki/Central_Zonal_Council}
}
\examples{
plot_map(include = .central, labels = TRUE)
}
\keyword{datasets}
