% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\docType{data}
\name{.east}
\alias{.east}
\title{Eastern Zonal Council}
\format{
An object of class \code{character} of length 4.
}
\usage{
.east
}
\description{
Eastern Zonal Council comprises of Bihar, Jharkhand, Odisha, and West Bengal.
}
\details{
See
\url{https://en.wikipedia.org/wiki/Eastern_Zonal_Council}
}
\examples{
plot_map(include = .east, labels = TRUE)
}
\keyword{datasets}
