% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\docType{data}
\name{.north}
\alias{.north}
\title{Northern Zonal Council}
\format{
An object of class \code{character} of length 8.
}
\usage{
.north
}
\description{
Northern Zonal Council comprises of Chandigarh, Delhi, Haryana, Himachal
Pradesh, Jammu and Kashmir, Ladakh, Punjab, and Rajasthan.
}
\details{
See
\url{https://en.wikipedia.org/wiki/Northern_Zonal_Council}
}
\examples{
plot_map(include = .north, labels = TRUE)
}
\keyword{datasets}
