% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\docType{data}
\name{.northeast}
\alias{.northeast}
\title{Northestern Zonal Council}
\format{
An object of class \code{character} of length 7.
}
\usage{
.northeast
}
\description{
North Eastern Council comprises of Assam, Arunachal Pradesh, Manipur,
Meghalaya, Mizoram, Nagaland and Tripura.
}
\details{
See
\url{https://en.wikipedia.org/wiki/North_Eastern_Council}
}
\examples{
plot_map(include = .northeast, labels = TRUE)
}
\keyword{datasets}
