% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\docType{data}
\name{.south}
\alias{.south}
\title{Southern Zonal Council (w/o Special invitees)}
\format{
An object of class \code{character} of length 6.
}
\usage{
.south
}
\description{
Southern Zonal Council comprises of Andhra Pradesh, Karnataka, Kerala,
Puducherry, Tamil Nadu, and Telangana.
}
\details{
See
\url{https://en.wikipedia.org/wiki/Southern_Zonal_Council}
}
\examples{
plot_map(include = .south, labels = TRUE)
}
\keyword{datasets}
