% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\docType{data}
\name{.southsp}
\alias{.southsp}
\title{Southern Zonal Council (w/ Special invitees)}
\format{
An object of class \code{character} of length 8.
}
\usage{
.southsp
}
\description{
Southern Zonal Council comprises of Andhra Pradesh, Karnataka, Kerala,
Puducherry, Tamil Nadu, and Telangana. Andaman and Nicobar and Lakshadweep
are special invitees of the Southern Zonal Council.
}
\details{
See
\url{https://en.wikipedia.org/wiki/Southern_Zonal_Council}
}
\examples{
plot_map(include = .southsp, labels = TRUE)
}
\keyword{datasets}
