% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\docType{data}
\name{.west}
\alias{.west}
\title{Western Zonal Council}
\format{
An object of class \code{character} of length 5.
}
\usage{
.west
}
\description{
Western Zonal Council comprises of Dadra and Nagar Haveli and Daman and Diu,
Goa, Gujarat, and Maharashtra.
}
\details{
See
\url{https://en.wikipedia.org/wiki/Western_Zonal_Council}
}
\examples{
plot_map(include = .west, labels = TRUE)
}
\keyword{datasets}
