% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{get_code_info}
\alias{get_code_info}
\title{Gets code11 info for either states or districts depending on input.
Helper function for S3 method [code_info()].}
\usage{
get_code_info(code, sortAndRemoveDuplicates)
}
\description{
Gets code11 info for either states or districts depending on input.
Helper function for S3 method [code_info()].
}
\keyword{internal}
