% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapindia-package.R
\docType{package}
\name{mapindia-package}
\alias{mapindia}
\alias{mapindia-package}
\title{mapindia: Plot Map of the Indian Subcontinent}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Get map data frames for the Indian subcontinent with different region levels (e.g., district, state). The package also offers convenience functions for plotting choropleths, visualizing spatial data, and handling state/district codes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shubhamdutta26/mapindia}
  \item \url{https://www.shubhamdutta.com/mapindia/}
  \item Report bugs at \url{https://github.com/shubhamdutta26/mapindia/issues}
}

}
\author{
\strong{Maintainer}: Shubham Dutta \email{shubhamdutta26@gmail.com} (\href{https://orcid.org/0000-0001-8484-0717}{ORCID}) [copyright holder]

}
\keyword{internal}
