% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statepop}
\alias{statepop}
\title{Indian Population (census and projections) by states}
\format{
A data frame with 36 rows and 8 variables.
}
\usage{
data(statepop)
}
\description{
Indian Population (census and projections) by states. \cr\cr
  The data is formatted for easy merging with output from [mapindia::map_india()].
}
\details{
\itemize{
  \item \code{code11} The 2-digit state code.
  \item \code{abbr} The 2-letter state abbreviation.
  \item \code{state} The full state name.
  \item \code{pop_1901} The 1901 population estimate (in number of people).
  \item \code{pop_1951} The 1951 population estimate (in number of people).
  \item \code{pop_2011} The 2011 population estimate (in number of people).
  \item \code{pop_2023} The 2023 population estimate (in number of people).
  \item \code{pop_2024} The 2024 population estimate (in number of people).
}
}
\keyword{data}
