% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{theme_map}
\alias{theme_map}
\title{Convenient theme map}
\usage{
theme_map(base_size = 9, base_family = "")
}
\description{
This creates a nice map theme for use in [plot_map()].
It originated from the `ggthemes` package located at this repository:
  \url{https://github.com/jrnold/ggthemes}.

This function was manually rewritten here to avoid the need for
 another package import.
}
\keyword{internal}
