% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_2011}
\alias{wb_2011}
\title{West Bengal population, sex-ratio, and literacy data (2011)}
\format{
A data frame with 23 rows and 8 variables.
}
\usage{
data(wb_2011)
}
\description{
West Bengal population, sex-ratio, and literacy data for 2011. \cr\cr
  The data is formatted for easy merging with output from [mapindia::map_india()].
}
\details{
\itemize{
  \item \code{code11} The 5-digit code corresponding to the district.
  \item \code{abbr} The 2-letter state abbreviation.
  \item \code{district} The full district name.
  \item \code{pop_2011} The 2011 population estimate (in number of people).
  \item \code{pop_increase_2011} The 2011 population increase (percent of people).
  \item \code{sex_ratio_2011} The 2011 sex ratio (females per 100 males).
  \item \code{literacy_per_2011} The 2011 literacy percentage.
  \item \code{pop_2011} The 2011 population density (unknown units).
}
}
\keyword{data}
